+++ tests/e2e/core-tests/CHANGELOG.md
@@ -27,6 +27,7 @@
 - Merchant Settings Shipping Zones
 - Shopper Variable product info updates on different variations
 - Merchant order emails flow
+- Merchant import products via CSV
 
 ## Fixed